+++ Movie.API/Domain/Services/MovieRepository.cs
@@ -14,7 +14,7 @@ namespace Movie.API.Domain.Services
         public MovieRepository(MoviesContext context) => _context = context;
 
         public IEnumerable<Movies> GetMovies() => _context.Movies;
-        public Movies GetMovie(int movieId) => _context.Movies.Where(x => x.Id == movieId).SingleOrDefault();
+        public Movies GetMovie(long movieId) => _context.Movies.Where(x => x.Id == movieId).SingleOrDefault();
 
         public void SaveDetail(Movies movie)
         {
